+++ extensions/iceworks-time-master/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - refactor: internal optimization
 - feat: editor info
+- fix: storage limit
 
 ## 0.3.6