+++ extensions/iceworks-time-master/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.3.7
+
+- refactor: internal optimization
+- feat: editor info
+- fix: storage limit
+
 ## 0.3.6
 
 - refactor: control views by configuration